Move into support/keypairgen
