+++ support/keypairgen/keypairgen_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 
 	"github.com/stellar/go/keypair"
-	"github.com/stellar/go/keypairgen"
 	"github.com/stellar/go/keypairgen/keypairgentest"
+	"github.com/stellar/go/keypairgen/keypsupport/keypairgen"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
